requirements.txt
Flask
gTTS
deep-translator
playsound
Werkzeug
SpeechRecognition
mistralai
python-dotenv
google-genai 
opencv-python 
pyaudio 
pillow 
mss